Testing: moving to Summit to get pickle files
