+++ AnalyzeAMSOS/Plots.py
@@ -16,6 +16,18 @@ def getSylinderPropertyNorm( frames, quant):
         propAll[i,:] = [np.linalg.norm( getattr( frame.sylinders[i], quant) ) for frame in frames]
     return propAll 
 
+def getSylinderMeanPropertyNorm( frames, quant):
+    # input: frames is a list of objects of class Frame
+    
+    propAll = getSylinderPropertyNorm( frames, quant)
+    shape = propAll.shape()
+
+    times = timestep*np.arange( len(shape[0]) )
+    means = np.mean( quantArr, axis=0)
+    stds = np.std( quantArr, axis=0)
+
+    return means, stds
+
 def getPropertyLabel( quant):
     # Get label with units
     if not ud.ud[quant]:
@@ -27,7 +39,7 @@ def getPropertyLabel( quant):
 def graphSylinderAllPropertyNorm( frames, quant, ax, timestep=1, alpha=1):
     # input: frames is a list of objects of class Frame
     
-    propAll = getSylinderPropertyNorm()
+    propAll = getSylinderPropertyNorm( frames, quant)
     for row in propAll:
         ax.plot( row, timestep*np.arange(len(row)), alpha=alpha)
 
@@ -37,12 +49,8 @@ def graphSylinderAllPropertyNorm( frames, quant, ax, timestep=1, alpha=1):
 def plotSylinderMeanPropertyNorm( frames, quant, ax, color='m', label='', timestep=1, err_alpha=0.3):
     # input: frames is a list of objects of class Frame
    
-    propAll = getSylinderPropertyNorm()
-    shape = propAll.shape()
-
-    times = timestep*np.arange( len(shape[0]) )
-    means = np.mean( quantArr, axis=0)
-    stds = np.std( quantArr, axis=0)
+    means, stds = getSylinderMeanPropertyNorm( frames, quant)
+    times = timestep*np.arange( len(means) )
 
     ax.plot(times, means, color=color, label=label)
     ax.fill_between(times, means-stds, means+stds, color=col, alpha=err_alpha)
